added xml comments

+++ Task6/Model/ExcelReportsModels/ExamResults.cs
@@ -7,10 +7,19 @@ using System.Threading.Tasks;
 
 namespace Model.ExcelReportsModels
 {
+    /// <summary>
+    /// Class ExamResults.
+    /// Implements the <see cref="Model.ExcelReportsModels.GroupsSessionResults" />
+    /// </summary>
+    /// <seealso cref="Model.ExcelReportsModels.GroupsSessionResults" />
     [Table(Name = "SessionResults")]
     public class ExamResults : GroupsSessionResults
     {
+        /// <summary>
+        /// Gets or sets the mark.
+        /// </summary>
+        /// <value>The mark.</value>
         [Column(Name = "Mark")]
-        public int Mark { get; set; }
+        public double Mark { get; set; }
     }
 }
